download breze kit and apply soft deleted , and  local , global scope in classromm  '

+++ app/Models/Classroom.php
@@ -2,13 +2,15 @@
 
 namespace App\Models;
 
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
-
 class Classroom extends Model
 {
-    use HasFactory;
+    use HasFactory,SoftDeletes;
 
     public static string $disk='public';
 
@@ -30,8 +32,40 @@ class Classroom extends Model
         ]);
 
     }
+    protected static function booted()
+    {
+        static::addGlobalScope('user',function (Builder $query){
+            $query->where('user_id','=',Auth::id());
+        });
+
+
+        // I can use withoutGlobalScope({name of global scope }) or withoutGlobalScopes() to avoid  scopes
+
+    }
+
     public static function deleteCoverImage($path)
     {
-       return Storage::disk(static::$disk)->delete($path);
+        if ($path && Storage::disk(static::$disk)->exists($path)){
+            return Storage::disk(static::$disk)->delete($path);
+
+        }
+    }
+
+
+
+    // local scopes
+
+    public function scopeActive(Builder $query)
+    {
+        $query->where('status','=','active');
+    }
+    public function scopeRecent(Builder $query)
+    {
+        $query->orderBy('updated_at','DESC');
+    }
+    public function scopeStatus(Builder $query, String $status = 'active')
+    {
+        $query->where('status','=',$status);
+
     }
 }
